Refactor: 타입 리팩토링
요기요 모델의 타입 (int -> long) 수정

+++ JsoupCrawling/src/main/java/_3jeon/server/yogiyo/model/YMenu.java
@@ -11,12 +11,12 @@ import java.util.List;
 @AllArgsConstructor
 public class YMenu {
     boolean soldout;
-    int review_count;
+    long review_count;
     String subtitle;
     String description;
-    int price;
+    long price;
     String slug;
     List<YSubMenu> subMenuList;
-    int id;
+    long id;
     String name;
 }
